Fixed some bugs
Fixed some bugs in the Drawer lists' component as they isn't a unique key for each child.
Fixed some bugs in the quranPlayerState context as it takes the initialState that is computed for one time for each load so to take the same state from the localStirage we must parse it each time the component mount.